bigclive.sh: add support for subdirectories for different video types
Significant re-write to add support for subdirectories:
announcements - channel announcements
tattoo - tattoo related behind-the-scenes
bigclivelive - livestreams

On its own, ls -1 can't really traverse subdirectories in a useful way
that preserves the ordering of the timestamps.
The naïve approach to fix this had significant performance issues when
searching the directories across a CIFS network share.
Changed the script to use tree with several advanced options for
filtering and output format.