Add binary_merge.py
A script for creating a merged, consensus output from multiple input
binary files.
Primary use case is to repair large binary files by comparing several
degraded copies.
No alignment is attempted - input files are all expected to be the
"same" file, potentially with "corrupted" spots or gaps that are
filled with 00 bytes.

A problem with simple consensus is that it could very easily be the
case that several input files have 00 fill for unreadable sectors
while a minority have the correct bytes. In this situation, 00 could
be taken as the majority. So, there is a command line option to
ignore 00 bytes for the purposes of consensus determination.

Another problem with simple consensus is that there may be no
majority or a tie for majority. The default behavior in these
cases is to output a 00 byte. A command line option exists to
instead use the contents of the first input file.