uri-open.sh: podcast: support URI with same name
If the resulting basename of the file to be downloaded is common
(e.g. all podcasts in a series are called "media.mp3"), then the old
behavior was to consider them the same file and attempt to play the
podcast rather than downloading the new one (and potentially over-
writing the old one).

Prepend the filename with the hash of the URI (actually 16-bit CRC)
and use that as the basis of the check for whether it is already
downloaded.